Reduced memory usage

+++ Program.cs
@@ -7,11 +7,11 @@ namespace Shanks
 {
 	public static class Program
 	{
-		private static BigInteger modulus;      // Max 318 numbers
-		private static BigInteger @base;        // Max 318 numbers
-		private static BigInteger order;        // Max 60 numbers
-		private static int        numberOfKeys; // Max 6 numbers
-		private static BigInteger bigStepExponent;
+		private static BigInteger modulus; // Max 318 numbers
+		private static BigInteger @base;   // Max 318 numbers
+		private static long order;         // Max 60 numbers
+		private static int numberOfKeys;   // Max 6 numbers
+		private static long bigStepExponent;
 		private static BigInteger bigStepValue;
 
 		public static void Main(string[] args)
@@ -19,7 +19,7 @@ namespace Shanks
 			char[] legalCharacters = { '0', '1', '2', '3', '4', '5', '6', '7', '8', '9' };
 			modulus         = BigInteger.Parse(Console.ReadLine().TakeWhile(c => legalCharacters.Contains(c)).ToArray());
 			@base           = BigInteger.Parse(Console.ReadLine().TakeWhile(c => legalCharacters.Contains(c)).ToArray());
-			order           = BigInteger.Parse(Console.ReadLine().TakeWhile(c => legalCharacters.Contains(c)).ToArray());
+			order           = long.Parse(Console.ReadLine()      .TakeWhile(c => legalCharacters.Contains(c)).ToArray());
 			numberOfKeys    = int.Parse(Console.ReadLine()       .TakeWhile(c => legalCharacters.Contains(c)).ToArray());
 			bigStepExponent = FastSquareRoot(ref order, ref numberOfKeys);
 			bigStepValue    = BigInteger.ModPow(@base, bigStepExponent, modulus);
@@ -38,11 +38,11 @@ namespace Shanks
 			}
 		}
 
-		static BigInteger FastSquareRoot(ref BigInteger orderRef, ref int numberOfKeysRef)
+		static long FastSquareRoot(ref long orderRef, ref int numberOfKeysRef)
 		{
-			BigInteger bigInteger = orderRef / numberOfKeysRef;
-			int        bits = 0;
-			BigInteger i    = 1;
+			long bigInteger = orderRef / numberOfKeysRef;
+			int  bits       = 0;
+			long i          = 1;
 			while (i << bits < bigInteger) { bits += 1; }
 			return bigInteger >> (bits / 2);
 		}
@@ -53,7 +53,7 @@ namespace Shanks
 
 			index.Add(1, 0);
 			BigInteger power = 1;
-			long bigSteps = (long) (order / bigStepExponent);
+			long bigSteps = order / bigStepExponent;
 			for (BigInteger i = 1; i < bigSteps + 1; i++)
 				index.Add(power = power * bigStepValue % modulus, (long) (i * bigStepExponent));
 
@@ -62,7 +62,7 @@ namespace Shanks
 
 		static void DiscreteLog(ref BigInteger x, ref Dictionary<BigInteger, long> index)
 		{
-			for (BigInteger smallSteps = 0; smallSteps < bigStepExponent; smallSteps++, x = x * @base % modulus)
+			for (long smallSteps = 0; smallSteps < bigStepExponent; smallSteps++, x = x * @base % modulus)
 				if (index.TryGetValue(x, out var exponent)) {
 					var result = exponent - smallSteps;
 					x = result >= 0 ? result : order - smallSteps;
